+++ inc/lilyParser/lilyInterpreter.ts
@@ -4,6 +4,7 @@ import {WHOLE_DURATION_MAGNITUDE, lcmMulti, lcm} from "./utils";
 import {parseRaw, getDurationSubdivider} from "./lilyTerms";
 import LogRecorder from "../logRecorder";
 import {StaffContext, PitchContextTable} from "../pitchContext";
+import * as idioms from "./idioms";
 
 import {
 	// eslint-disable-next-line
@@ -29,15 +30,6 @@ type MusicListener = (music: BaseTerm, context: TrackContext) => void;
 
 
 type ContextDict = {[key: string]: string};
-
-
-/*interface NoteRef {
-	track?: number;
-	event: MusicEvent;
-	pitch: ChordElement;
-	ctx: PitchContext;
-	staffName: string;
-};*/
 
 
 interface PitchContextTerm {
@@ -74,12 +66,21 @@ class LilyStaffContext extends StaffContext {
 
 		if (term.clef)
 			this.setClef(term.clef.y, term.clef.value);
-
 		if (Number.isFinite(term.octaveShift))
-			this.setOctaveShift(this.octaveShift);
-
+			this.setOctaveShift(-term.octaveShift);
 		if (Number.isFinite(term.key)) {
-			// TODO:
+			this.resetKeyAlters();
+
+			if (term.key) {
+				const step = term.key > 0 ? 1 : -1;
+				for (let p = step; p / term.key <= 1; p += step) {
+					const index = ((step > 0 ? p - 1 : p) + 70) % 7;
+					const note = idioms.PHONETS.indexOf(idioms.FIFTH_PHONETS[index]);
+
+					this.keyAlters[note] = (this.keyAlters[note] || 0) + step;
+				}
+				this.dirty = true;
+			}
 		}
 
 		if (term.pitches) {
@@ -300,16 +301,6 @@ export class MusicTrack {
 			getCurrentTerm(track.staffName).tick = track.tick;
 
 			if (term instanceof Chord) {
-				/*//const index = context.snapshot({tick: term._tick});
-				//const ctx = context.track.contexts[index];
-
-				term.pitchElements.forEach(pitch => notes.push({
-					event: term,
-					pitch,
-					//ctx,
-					staffName: track.staffName,
-				}));*/
-
 				const pcTerm = getCurrentTerm(track.staffName);
 				pcTerm.event = term;
 				pcTerm.pitches = term.pitchElements;
@@ -336,12 +327,10 @@ export class MusicTrack {
 					break;
 				}
 			}
-			else if (term instanceof KeySignature) {
-				// TODO:
-			}
-			else if (term instanceof OctaveShift) {
-				// TODO:
-			}
+			else if (term instanceof KeySignature)
+				getCurrentTerm(track.staffName).key = term.key;
+			else if (term instanceof OctaveShift)
+				getCurrentTerm(track.staffName).octaveShift = term.value;
 		};
 		new TrackContext(this, {listener}).execute(this.music);
 
@@ -844,61 +833,6 @@ export default class LilyInterpreter {
 
 
 	getNotation ({logger = new LogRecorder()} = {}): LilyNotation.Notation {
-		/*const pitchContexts = this.staffNames.map(name => ({
-			name,
-			items: [],
-		}));
-
-		const tracks: NoteRef[][] = this.musicTracks.map((track, i) => {
-			const {notes, group} = track.generateStaffTracks({logger});
-			//console.log("group:", group);
-
-			pitchContexts.forEach(({name, items}) => {
-				const track = group[name];
-				if (track) {
-					items.push(...track.contexts.map(context => ({
-						tick: context.tick,
-						context,
-					})));
-				}
-			});
-
-			return notes.map(note => ({track: i, ...note}));
-		});
-		const noteRefs = [].concat(...tracks);
-
-		const pitchContextGroup = pitchContexts.map(({items}) => {
-			items.sort((i1, i2) => i1.tick - i2.tick);
-			items.forEach((item, i) => {
-				item.endTick = (i + 1 < items.length ? items[i + 1].tick : Infinity);
-				item.context.index = i;
-			});
-
-			return new PitchContextTable({items});
-		});
-
-		const staffNamesIndices = this.staffNames.reduce((indices, name, index) => ({...indices, [name]: index}), {});
-
-		const notes = noteRefs.map(({event, pitch, staffName}) => ({
-			channel: 0,
-			start: event._tick,
-			duration: event.durationMagnitude,
-			startTick: event._tick,
-			endTick: event._tick + event.durationMagnitude,
-			pitch: pitch.absolutePitchValue + (pitch._transposition || 0),
-			velocity: 127,
-			id: pitch.href,
-			tied: pitch._tied,
-			rest: event.isRest,
-			staffTrack: staffNamesIndices[staffName],
-			// TODO: snapshot staff context every term tick, bind contextIndex by tick
-			//contextIndex: ctx.index,
-		})).sort((n1, n2) => n1.startTick - n2.startTick);
-
-		return {
-			notes,
-			pitchContextGroup,
-		};*/
 		const pcTerms = [].concat(...this.musicTracks.map((track, i) => track.generateStaffTracks().map(term => ({track: i, ...term}))));
 		//console.log("pcTerms:", pcTerms);